backend/requirements.txt
# ChainMind Backend — Python Dependencies
# ────────────────────────────────────────

# ── Core Web Framework ─────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# ── Database ───────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
# psycopg2-binary==2.9.10
# alembic==1.14.0

# ── Redis ──────────────────────────────────────────────────────────
redis[asyncio]==5.2.0

# ── HTTP / API Clients ─────────────────────────────────────────────
httpx==0.27.2
aiohttp==3.11.7
requests==2.32.3
beautifulsoup4==4.12.3

# ── Data Engineering ───────────────────────────────────────────────
pandas==2.2.3
numpy==1.26.4
pyarrow==18.0.0
scipy==1.14.1
scikit-learn==1.5.2

# ── Statistical Forecasting ────────────────────────────────────────
statsmodels==0.14.4

# ── ML / Deep Learning ─────────────────────────────────────────────
torch==2.5.1
# pytorch-forecasting==1.1.1   # Install separately if GPU needed
# pytorch-lightning==2.4.0

# ── Reinforcement Learning ─────────────────────────────────────────
stable-baselines3==2.3.2
gymnasium==0.29.1
shimmy

# ── NLP ────────────────────────────────────────────────────────────
# transformers==4.46.3         # Install for local HuggingFace models
# sentence-transformers==3.3.1

# ── Optimization ───────────────────────────────────────────────────
ortools==9.11.4210
pulp==2.9.0

# ── LangChain / LLM ────────────────────────────────────────────────
google-generativeai==0.8.3
langchain==0.3.9
langchain-google-genai==2.0.4

# ── MLflow Experiment Tracking ─────────────────────────────────────
mlflow==2.18.0

# ── Scheduling ─────────────────────────────────────────────────────
apscheduler==3.10.4

# ── Utilities ──────────────────────────────────────────────────────
python-dateutil==2.9.0
tqdm==4.67.1
loguru==0.7.2
